Formatting from 100 line limit to 120

+++ program/include/timer.hpp
@@ -12,7 +12,7 @@ namespace cse::time
     ~Timer();
 
   private:
-    float microseconds() const;
+    long long microseconds() const;
     float milliseconds() const;
     float seconds() const;
