+++ program/source/timer.cpp
@@ -6,18 +6,14 @@
 
 namespace cse::time
 {
-  Timer::Timer(const std::string &i_name) : name(i_name)
-  {
-    time = std::chrono::high_resolution_clock::now();
-  }
+  Timer::Timer(const std::string &i_name) : name(i_name) { time = std::chrono::high_resolution_clock::now(); }
   Timer::~Timer() { std::cout << name << ": " << milliseconds() << "ms\n"; }
 
-  float Timer::microseconds() const
+  long long Timer::microseconds() const
   {
-    return (float)std::chrono::duration_cast<std::chrono::microseconds>(
-             std::chrono::high_resolution_clock::now() - time)
+    return std::chrono::duration_cast<std::chrono::microseconds>(std::chrono::high_resolution_clock::now() - time)
       .count();
   }
-  float Timer::milliseconds() const { return microseconds() * 0.001f; }
+  float Timer::milliseconds() const { return (float)microseconds() * 0.001f; }
   float Timer::seconds() const { return milliseconds() * 0.001f; }
 }
